pyproject.toml
[project]
name = "voiceapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "piper-tts>=1.4.1",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
